Chore: Move lib/report-translator to lib/util/report-translator
